services/sensor-simulator/requirements.txt
paho-mqtt==2.1.0
pandas==3.0.2
boto3==1.42.89
python-dotenv==1.2.2